Fixed scraping bug for the medal table
Scrolling wasn't occuring properly during scraping of medal table from
olympics site. We appeared to be just jumping to the end of the webpage,
which is because we weren't properly looking at the current pageYOffset.
This change also attempts to make things slightly more rebust by
incrementally scrolling in viewport//5 steps and sleeping.